requirements.txt
google-re2==1.1
googleapis-common-protos==1.60.0
protobuf==4.24.2
apache-airflow
cassandra-driver
Flask
pyspark
kafka-python
pydantic